Do not generate separate items for parent products (#11)
