+++ src/FeedContext/Google/Shopping/ProductItemContext.php
@@ -27,10 +27,12 @@ use Sylius\Component\Core\Model\ImagesAwareInterface;
 use Sylius\Component\Core\Model\ProductInterface;
 use Sylius\Component\Core\Model\ProductTranslationInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
+use Sylius\Component\Inventory\Checker\AvailabilityCheckerInterface;
 use Sylius\Component\Locale\Model\LocaleInterface;
 use Sylius\Component\Product\Resolver\ProductVariantResolverInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\RouterInterface;
+use Webmozart\Assert\Assert;
 
 class ProductItemContext implements ItemContextInterface
 {
@@ -40,27 +42,27 @@ class ProductItemContext implements ItemContextInterface
     /** @var CacheManager */
     private $cacheManager;
 
-    /** @var ItemContextInterface */
-    private $productVariantItemContext;
-
     /** @var ProductVariantPriceCalculatorInterface */
     private $productVariantPriceCalculator;
 
     /** @var ProductVariantResolverInterface */
     private $productVariantResolver;
 
+    /** @var AvailabilityCheckerInterface */
+    private $availabilityChecker;
+
     public function __construct(
         RouterInterface $router,
         CacheManager $cacheManager,
         ProductVariantResolverInterface $productVariantResolver,
         ProductVariantPriceCalculatorInterface $productVariantPriceCalculator,
-        ItemContextInterface $productVariantItemContext
+        AvailabilityCheckerInterface $availabilityChecker
     ) {
         $this->router = $router;
         $this->cacheManager = $cacheManager;
-        $this->productVariantItemContext = $productVariantItemContext;
         $this->productVariantPriceCalculator = $productVariantPriceCalculator;
         $this->productVariantResolver = $productVariantResolver;
+        $this->availabilityChecker = $availabilityChecker;
     }
 
     /**
@@ -76,50 +78,58 @@ class ProductItemContext implements ItemContextInterface
         }
 
         $translation = $this->getTranslation($product, $locale->getCode());
-
-        $data = new Product();
-        $data->setId($product->getCode());
-        $data->setImageLink($this->getImageLink($product));
-        $data->setAvailability(Availability::outOfStock());
-        $data->setPrice($this->getPrice($product, $channel));
-
-        if (null !== $translation) {
-            $data->setTitle($translation->getName());
-            $data->setDescription($translation->getDescription());
-            $data->setLink($this->getLink($locale, $translation));
-        }
-
-        $data->setCondition($product instanceof ConditionAwareInterface ? Condition::fromValue($product->getCondition()) : Condition::new());
-        $data->setItemGroupId($product->getCode());
-
-        if ($product instanceof BrandAwareInterface && $product->getBrand() !== null) {
-            $data->setBrand((string) $product->getBrand());
-        }
-
-        if ($product instanceof GtinAwareInterface && $product->getGtin() !== null) {
-            $data->setGtin((string) $product->getGtin());
-        }
-
-        if ($product instanceof MpnAwareInterface && $product->getMpn() !== null) {
-            $data->setMpn((string) $product->getMpn());
-        }
-
-        if ($product instanceof SizeAwareInterface && $product->getSize() !== null) {
-            $data->setSize((string) $product->getSize());
-        }
-
-        if ($product instanceof ColorAwareInterface && $product->getColor() !== null) {
-            $data->setColor((string) $product->getColor());
-        }
-
         $contextList = new ContextList();
-        $contextList->add($data);
-
         foreach ($product->getVariants() as $variant) {
-            $variantContextList = $this->productVariantItemContext->getContextList($variant, $channel, $locale);
-            foreach ($variantContextList as $item) {
-                $contextList->add($item);
+            Assert::isInstanceOf($variant, ProductVariantInterface::class);
+            $data = new Product();
+            $data->setId($variant->getCode());
+            $data->setItemGroupId($product->getCode());
+            $data->setImageLink($this->getImageLink($product));
+            $data->setAvailability($this->getAvailability($variant));
+
+            [$price, $salePrice] = $this->getPrices($variant, $channel);
+            $data->setPrice($price);
+            $data->setSalePrice($salePrice);
+
+            if (null !== $translation) {
+                $data->setTitle($translation->getName());
+                $data->setDescription($translation->getDescription());
+                $data->setLink($this->getLink($locale, $translation));
             }
+
+            $data->setCondition($product instanceof ConditionAwareInterface ? Condition::fromValue($product->getCondition()) : Condition::new());
+
+            if ($variant instanceof BrandAwareInterface && $variant->getBrand() !== null) {
+                $data->setBrand((string) $variant->getBrand());
+            } elseif ($product instanceof BrandAwareInterface && $product->getBrand() !== null) {
+                $data->setBrand((string) $product->getBrand());
+            }
+
+            if ($variant instanceof GtinAwareInterface && $variant->getGtin() !== null) {
+                $data->setGtin((string) $variant->getGtin());
+            } elseif ($product instanceof GtinAwareInterface && $product->getGtin() !== null) {
+                $data->setGtin((string) $product->getGtin());
+            }
+
+            if ($variant instanceof MpnAwareInterface && $variant->getMpn() !== null) {
+                $data->setMpn((string) $variant->getMpn());
+            } elseif ($product instanceof MpnAwareInterface && $product->getMpn() !== null) {
+                $data->setMpn((string) $product->getMpn());
+            }
+
+            if ($variant instanceof SizeAwareInterface && $variant->getSize() !== null) {
+                $data->setSize((string) $variant->getSize());
+            } elseif ($product instanceof SizeAwareInterface && $product->getSize() !== null) {
+                $data->setSize((string) $product->getSize());
+            }
+
+            if ($variant instanceof ColorAwareInterface && $variant->getColor() !== null) {
+                $data->setColor((string) $variant->getColor());
+            } elseif ($product instanceof ColorAwareInterface && $product->getColor() !== null) {
+                $data->setColor((string) $product->getColor());
+            }
+
+            $contextList->add($data);
         }
 
         return $contextList;
@@ -164,15 +174,39 @@ class ProductItemContext implements ItemContextInterface
         return $this->cacheManager->getBrowserPath($images->first()->getPath(), 'sylius_shop_product_large_thumbnail');
     }
 
-    private function getPrice(ProductInterface $product, ChannelInterface $channel): ?Price
+    /**
+     * Index 0 equals the price
+     * Index 1 equals the sale price (if set)
+     */
+    private function getPrices(ProductVariantInterface $variant, ChannelInterface $channel): array
     {
-        /** @var ProductVariantInterface|null $variant */
-        $variant = $this->productVariantResolver->getVariant($product);
-        if (null === $variant) {
-            return null;
+        $channelPricing = $variant->getChannelPricingForChannel($channel);
+
+        if (null === $channelPricing) {
+            return [null, null];
         }
 
-        $price = $this->productVariantPriceCalculator->calculate($variant, ['channel' => $channel]);
+        $originalPrice = $channelPricing->getOriginalPrice();
+        $price = $channelPricing->getPrice();
+
+        if (null === $price) {
+            return [null, null];
+        }
+
+        if (null === $originalPrice) {
+            return [$this->createPrice($price, $channel), null];
+        }
+
+        return [$this->createPrice($originalPrice, $channel), $this->createPrice($price, $channel)];
+    }
+
+    private function getAvailability(ProductVariantInterface $product): Availability
+    {
+        return $this->availabilityChecker->isStockAvailable($product) ? Availability::inStock() : Availability::outOfStock();
+    }
+
+    private function createPrice(int $price, ChannelInterface $channel): ?Price
+    {
         $baseCurrency = $channel->getBaseCurrency();
         if (null === $baseCurrency) {
             return null;
