+++ src/Resources/config/services/feed_context.xml
@@ -9,13 +9,7 @@
             <argument type="service" id="liip_imagine.cache.manager"/>
             <argument type="service" id="sylius.product_variant_resolver.default"/>
             <argument type="service" id="sylius.calculator.product_variant_price"/>
-            <argument type="service" id="setono_sylius_feed.feed_context.google.shopping.product_variant_item"/>
-        </service>
-
-        <service id="setono_sylius_feed.feed_context.google.shopping.product_variant_item" class="Setono\SyliusFeedPlugin\FeedContext\Google\Shopping\ProductVariantItemContext">
             <argument type="service" id="sylius.availability_checker"/>
-            <argument type="service" id="liip_imagine.cache.manager"/>
-            <argument type="service" id="router"/>
         </service>
     </services>
 </container>
